BUILD : 1.7.5 (367) BUGS : NOTES : Changed UserKeys from uint to long uint.
git-svn-id: svn://svn.anope.org/anope/trunk@367 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@238 5417fbe8-f217-4b02-8779-1006273d7864

+++ include/extern.h
@@ -498,9 +498,9 @@ E char *DefconMessage;
 E char *DefConAkillReason;
 E char *DefConOffMessage;
 
-E unsigned int UserKey1;
-E unsigned int UserKey2;     
-E unsigned int UserKey3;
+E long unsigned int UserKey1;
+E long unsigned int UserKey2;     
+E long unsigned int UserKey3;
 /**** converter.c ****/
 
 E int convert_ircservices_44(void);
